add utils for function, change diagnose db structure model (date->diagnosisDate, physicalActivityLevel -> physicalActivityInMinute, createdAt (new))

+++ server/controllers/diagnosis.js
@@ -1,35 +1,40 @@
 const User = require('../models/User');
 const Diagnosis = require('../models/Diagnosis');
-
-// Simple function to diagnose sleep disorder
-const diagnoseSleepDisorder = (data) => {
-  const { sleepDuration, qualityOfSleep, physicalActivityLevel, stressLevel, bloodPressure, heartRate, dailySteps } = data;
-  if (sleepDuration < 4 || sleepDuration > 10 || qualityOfSleep === 'poor' || stressLevel === 'high') {
-    return { disorder: 'Possible sleep disorder detected', solution: 'Consult a sleep specialist, maintain a regular sleep schedule, reduce stress.' };
-  } else {
-    return { disorder: 'Normal sleep patterns', solution: 'Continue maintaining healthy sleep habits.' };
-  }
-};
+const { 
+  diagnoseSleepDisorder, 
+  calculateBMI,
+  getCurrentDate,
+  getTimestamp,
+  formatDiagnosisDate
+} = require('../utils/Utils');
 
 exports.addDiagnose = async (req, res) => {
   try {
-    const { uid, date, gender, age, name, BMIcategory, sleepDuration, qualityOfSleep, physicalActivityLevel, stressLevel, bloodPressure, heartRate, dailySteps } = req.body;
+    const user = req.user;
+    const { sleepDuration, qualityOfSleep, physicalActivityInMinute, stressLevel, bloodPressure, heartRate, dailySteps, diagnosisDate, height, weight } = req.body;
 
+    // Use the formatDiagnosisDate function
+    const formattedDiagnosisDate = formatDiagnosisDate(diagnosisDate);
+
+    const BMI = calculateBMI(weight, height);
     const diagnosisData = {
-      uid,
-      date,
-      gender,
-      age,
-      name,
-      BMIcategory,
+      uid: user._id,
+      diagnosisDate: formattedDiagnosisDate,
+      gender: user.gender,
+      age: user.age,
+      name: user.name,
+      height,
+      weight,
+      BMIcategory: BMI,
       sleepDuration,
       qualityOfSleep,
-      physicalActivityLevel,
+      physicalActivityInMinute,
       stressLevel,
       bloodPressure,
       heartRate,
       dailySteps,
-      timestamp: new Date()
+      createdAt: getCurrentDate(),
+      timestamp: getTimestamp()
     };
 
     const { disorder, solution } = diagnoseSleepDisorder(diagnosisData);
@@ -39,12 +44,7 @@ exports.addDiagnose = async (req, res) => {
     const newDiagnosis = new Diagnosis(diagnosisData);
     await newDiagnosis.save();
 
-    let user = await User.findOne({ _id: uid });
-    if (!user) {
-      user = new User({ _id: uid, name, age, gender, diagnoses: [newDiagnosis._id] });
-    } else {
-      user.diagnoses.push(newDiagnosis._id);
-    }
+    user.diagnoses.push(newDiagnosis._id);
     await user.save();
 
     res.status(201).send({ message: 'Diagnosis complete', diagnosis: newDiagnosis });
@@ -53,9 +53,10 @@ exports.addDiagnose = async (req, res) => {
   }
 };
 
+
 exports.getAllDiagnosesByUser = async (req, res) => {
   try {
-    const uid = req.user._id;  // Assuming user ID is set in req.user by the auth middleware
+    const uid = req.user._id;  // User ID from authenticated user
     const diagnoses = await Diagnosis.find({ uid });
     res.status(200).send(diagnoses);
   } catch (err) {
@@ -67,6 +68,16 @@ exports.editDiagnosis = async (req, res) => {
   try {
     const { id } = req.params;
     const updatedData = req.body;
+
+    if (updatedData.weight && updatedData.height) {
+      updatedData.BMIcategory = calculateBMI(updatedData.weight, updatedData.height);
+    }
+
+    const { disorder, solution } = diagnoseSleepDisorder(updatedData);
+    updatedData.sleepDisorder = disorder;
+    updatedData.solution = solution;
+    updatedData.timestamp=getTimestamp();
+
     const updatedDiagnosis = await Diagnosis.findByIdAndUpdate(id, updatedData, { new: true });
     res.status(200).send({ message: 'Diagnosis updated', diagnosis: updatedDiagnosis });
   } catch (err) {
@@ -77,6 +88,10 @@ exports.editDiagnosis = async (req, res) => {
 exports.deleteDiagnosis = async (req, res) => {
   try {
     const { id } = req.params;
+    const diagnosis = await Diagnosis.findById(id);
+    if (!diagnosis) {
+      return res.status(404).send({ message: 'Diagnosis not found' });
+    }
     await Diagnosis.findByIdAndDelete(id);
     res.status(200).send({ message: 'Diagnosis deleted' });
   } catch (err) {
